feat: add structure for opening a url using the browser app
Create the Xcode project structure and eliminate the default structure.
Add the engine structure with the 'openExternalBrowser' method and the router protocol with 'openInSafari'. This allows the library's caller to add the behaviour to whoever will be responsible for the routing (AppDelegate).
Add the unit tests for the feature.

References: https://outsystemsrd.atlassian.net/browse/RMET-3421