feat: add events for SystemBrowser
The events are part of the OSIABSafariViewControllerRouterAdapter class. The same applies for the OSIABSystemBrowserOptions structure, to make things clearer.
Set the SFSafariViewControllerDelegate and UIAdaptivePresentationControllerDelegate delegates so that the events can be intercepted and the new events triggered.
Add unit tests to validate the feature.

References: https://outsystemsrd.atlassian.net/browse/RMET-3429