feat(ios): implement close
This is applied to the Router's concrete implementations. To achieve this, a specific subclass was created where the 'dismiss' method is override to return the 'onBrowserClosed' callback. With this, previously used methods that did similar were removed (safariViewControllerDidFinish).
Create method accelerators to transform OSIABWebViewOptions to OSIABWebViewConfigurationModel and OSIABWebViewUIModel.
Update unit tests.

References: https://outsystemsrd.atlassian.net/browse/RMET-3427