fix: add support for iOS 13
This is needed as Capacitor still supports this version of iOS.
Set SWIFT_VERSION with 5.0 at project level. This guarantees that all versions that Xcode 15 supports (all until 5.9) are guaranteed while building the xcframework.
The Continuous Integration pipeline is updated to use Xcode 15.0 instead of 15.1 as it's the minimum Xcode version supported by Capacitor.